requirements.txt
# Core dependencies
python-telegram-bot==21.5
aiohttp==3.10.2
requests==2.31.0
python-dotenv==1.0.1

# Database and async support
aiosqlite==0.19.0

# Security and validation
cryptography==42.0.8

# Development and testing
pytest==8.0.0
pytest-asyncio==0.23.0

# Security scanning tools (optional)
bandit==1.7.9
safety==3.2.7
pip-audit==2.6.1

# Code quality
ruff==0.3.0
mypy==1.8.0aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
aiosqlite==0.21.0
anyio==4.9.0
APScheduler==3.11.0
attrs==25.1.0
beautifulsoup4==4.13.3
bs4==0.0.2
cattrs==24.1.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
construct==2.10.68
construct-typing==0.5.6
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.3.0
fuzzywuzzy==0.18.0
greenlet==3.1.1
h11==0.14.0
helius-sdk==0.0.11
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
jsonalias==0.1.1
Levenshtein==0.27.1
multidict==6.3.0
numpy==2.2.3
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
platformdirs==4.3.6
playwright==1.50.0
polars==1.26.0
propcache==0.3.1
pyarrow==19.0.1
pyee==12.1.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.27.1
python-telegram-bot==22.0
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
requests-cache==1.2.1
requests-oauthlib==2.0.0
safetensors==0.5.3
selenium==4.29.0
six==1.17.0
sniffio==1.3.1
solana==0.36.6
solders==0.26.0
sortedcontainers==2.4.0
soupsieve==2.6
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.49.0
trio==0.29.0
trio-websocket==0.12.2
tweepy==4.15.0
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
uploadserver==6.0.0
url-normalize==1.4.3
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0
wsproto==1.2.0
yarl==1.18.3